+++ src/createPersistoid.ts
@@ -9,9 +9,7 @@ export default function createPersistoid(config: PersistConfig<any>): Persistoid
   const whitelist: string[] | null = config.whitelist || null
   const transforms = config.transforms || []
   const throttle = config.throttle || 0
-  const storageKey = `${
-    config.keyPrefix !== undefined ? config.keyPrefix : KEY_PREFIX
-  }${config.key}`
+  const storageKey = `${config.keyPrefix !== undefined ? config.keyPrefix : KEY_PREFIX}${config.key}`
   const storage = config.storage
   let serialize: (x: any) => any
   if (config.serialize === false) {
@@ -27,12 +25,12 @@ export default function createPersistoid(config: PersistConfig<any>): Persistoid
   let lastState: KeyAccessState = {}
   const stagedState: KeyAccessState = {}
   const keysToProcess: string[] = []
-  let timeIterator: any = null
+  let writeTimeout: any = null
   let writePromise: Promise<any> | null = null
 
   const update = (state: KeyAccessState) => {
     // add any changed keys to the queue
-    Object.keys(state).forEach(key => {
+    Object.keys(state).forEach((key) => {
       if (!passWhitelistBlacklist(key)) return // is keyspace ignored? noop
       if (lastState[key] === state[key]) return // value unchanged? noop
       if (keysToProcess.indexOf(key) !== -1) return // is key already queued? noop
@@ -41,7 +39,7 @@ export default function createPersistoid(config: PersistConfig<any>): Persistoid
 
     //if any key is missing in the new state which was present in the lastState,
     //add it for processing too
-    Object.keys(lastState).forEach(key => {
+    Object.keys(lastState).forEach((key) => {
       if (
         state[key] === undefined &&
         passWhitelistBlacklist(key) &&
@@ -52,18 +50,19 @@ export default function createPersistoid(config: PersistConfig<any>): Persistoid
       }
     })
 
-    // start the time iterator if not running (read: throttle)
-    if (timeIterator === null) {
-      timeIterator = setInterval(processNextKey, throttle)
-    }
-
     lastState = state
+
+    if (!throttle) {
+      while (keysToProcess.length) {
+        processNextKey()
+      }
+    } else if (writeTimeout === null) {
+      writeTimeout = setTimeout(flush, throttle)
+    }
   }
 
   function processNextKey() {
     if (keysToProcess.length === 0) {
-      if (timeIterator) clearInterval(timeIterator)
-      timeIterator = null
       return
     }
 
@@ -79,10 +78,7 @@ export default function createPersistoid(config: PersistConfig<any>): Persistoid
       try {
         stagedState[key] = serialize(endState)
       } catch (err) {
-        console.error(
-          'redux-persist/createPersistoid: error serializing state',
-          err
-        )
+        console.error('redux-persist/createPersistoid: error serializing state', err)
       }
     } else {
       //if the endState is undefined, no need to persist the existing serialized content
@@ -96,35 +92,57 @@ export default function createPersistoid(config: PersistConfig<any>): Persistoid
 
   function writeStagedState() {
     // cleanup any removed keys just before write.
-    Object.keys(stagedState).forEach(key => {
+    Object.keys(stagedState).forEach((key) => {
       if (lastState[key] === undefined) {
         delete stagedState[key]
       }
     })
 
-    writePromise = storage
-      .setItem(storageKey, serialize(stagedState))
-      .catch(onWriteFail)
+    let serialized: any
+    try {
+      serialized = serialize(stagedState)
+    } catch (err) {
+      onWriteFail(err)
+      return
+    }
+
+    writePromise = storage.setItem(storageKey, serialized).catch(onWriteFail)
   }
 
   function passWhitelistBlacklist(key: string) {
-    if (whitelist && whitelist.indexOf(key) === -1 && key !== '_persist')
-      return false
+    if (whitelist && whitelist.indexOf(key) === -1 && key !== '_persist') return false
     if (blacklist && blacklist.indexOf(key) !== -1) return false
     return true
   }
 
   function onWriteFail(err: any) {
-    // @TODO add fail handlers (typically storage full)
     if (writeFailHandler) writeFailHandler(err)
     if (err && process.env.NODE_ENV !== 'production') {
-      console.error('Error storing data', err)
+      if (err.name === 'QuotaExceededError' || err.name === 'NS_ERROR_DOM_QUOTA_REACHED') {
+        console.error(
+          'redux-persist/createPersistoid: storage quota exceeded. ' +
+            'State will not be persisted until storage is freed.',
+          err
+        )
+      } else if (err.name === 'SecurityError') {
+        console.error(
+          'redux-persist/createPersistoid: storage access denied. ' +
+            'State will not be persisted. This can occur in private browsing mode.',
+          err
+        )
+      } else {
+        console.error('redux-persist/createPersistoid: error storing data', err)
+      }
     }
   }
 
-  const flush = () => {
+  function flush() {
     while (keysToProcess.length !== 0) {
       processNextKey()
+    }
+    if (writeTimeout) {
+      clearTimeout(writeTimeout)
+      writeTimeout = null
     }
     return writePromise || Promise.resolve()
   }