+++ src/persistStore.ts
@@ -1,13 +1,9 @@
-import type {
-  Persistor,
-  PersistorOptions,
-  PersistorState,
-} from './types'
+import type { Persistor, PersistorOptions, PersistorState, RehydrateAction } from './types'
 
 import { AnyAction, createStore, Store } from 'redux'
 import { FLUSH, PAUSE, PERSIST, PURGE, REGISTER, REHYDRATE } from './constants'
 
-type BoostrappedCb = () => any;
+type BoostrappedCb = () => any
 
 const initialState: PersistorState = {
   registry: [],
@@ -29,26 +25,15 @@ const persistorReducer = (state = initialState, action: AnyAction) => {
 }
 
 interface OptionToTestObject {
-  [key: string]: any;
+  [key: string]: any
 }
 
-export default function persistStore(
-  store: Store,
-  options?: PersistorOptions,
-  cb?: BoostrappedCb
-): Persistor {
+export default function persistStore(store: Store, options?: PersistorOptions, cb?: BoostrappedCb): Persistor {
   // help catch incorrect usage of passing PersistConfig in as PersistorOptions
   if (process.env.NODE_ENV !== 'production') {
     const optionsToTest: OptionToTestObject = options || {}
-    const bannedKeys = [
-      'blacklist',
-      'whitelist',
-      'transforms',
-      'storage',
-      'keyPrefix',
-      'migrate',
-    ]
-    bannedKeys.forEach(k => {
+    const bannedKeys = ['blacklist', 'whitelist', 'transforms', 'storage', 'keyPrefix', 'migrate']
+    bannedKeys.forEach((k) => {
       if (optionsToTest[k])
         console.error(
           `redux-persist: invalid option passed to persistStore: "${k}". You may be incorrectly passing persistConfig into persistStore, whereas it should be passed into persistReducer.`
@@ -57,11 +42,8 @@ export default function persistStore(
   }
   let boostrappedCb = cb || false
 
-  const _pStore = createStore(
-    persistorReducer,
-    initialState,
-    options && options.enhancer ? options.enhancer : undefined
-  )
+  const _pStore: Store<PersistorState, AnyAction> = createStore(persistorReducer, initialState, options && options.enhancer ? options.enhancer : undefined)
+
   const register = (key: string) => {
     _pStore.dispatch({
       type: REGISTER,
@@ -70,7 +52,7 @@ export default function persistStore(
   }
 
   const rehydrate = (key: string, payload: Record<string, unknown>, err: any) => {
-    const rehydrateAction = {
+    const rehydrateAction: RehydrateAction = {
       type: REHYDRATE,
       payload,
       err,
@@ -79,7 +61,7 @@ export default function persistStore(
     // dispatch to `store` to rehydrate and `persistor` to track result
     store.dispatch(rehydrateAction)
     _pStore.dispatch(rehydrateAction)
-    if (typeof boostrappedCb === "function" && persistor.getState().bootstrapped) {
+    if (typeof boostrappedCb === 'function' && persistor.getState().bootstrapped) {
       boostrappedCb()
       boostrappedCb = false
     }
@@ -117,7 +99,7 @@ export default function persistStore(
     },
   }
 
-  if (!(options && options.manualPersist)){
+  if (!(options && options.manualPersist)) {
     persistor.persist()
   }