+++ tests/utils/find.ts
@@ -1,4 +1,7 @@
-export default (collection: Array<Record<string, any>>, predicate: Record<string, string>): Record<string, any> | null => {
+export default (
+  collection: Array<Record<string, any>>,
+  predicate: Record<string, string>
+): Record<string, any> | null => {
   let result: Record<string, any> | null = null
   collection.forEach((value: any) => {
     if (value.type && value.type === predicate.type) {